% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prediction_functions.R
\docType{package}
\name{prediction_googleAuthR}
\alias{prediction_googleAuthR}
\alias{prediction_googleAuthR-package}
\title{Prediction API
Lets you access a cloud hosted machine learning service that makes it easy to build smart apps}
\description{
Auto-generated code by googleAuthR::gar_create_api_skeleton
 at 2017-03-05 20:06:18
filename: /Users/mark/dev/R/autoGoogleAPI/googlepredictionv14.auto/R/prediction_functions.R
api_json: api_json
}
\details{
Authentication scopes used are:
\itemize{
  \item https://www.googleapis.com/auth/devstorage.full_control
\item https://www.googleapis.com/auth/devstorage.read_only
\item https://www.googleapis.com/auth/devstorage.read_write
\item https://www.googleapis.com/auth/prediction
}
}
