% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prediction_functions.R
\name{trainedmodels.predict}
\alias{trainedmodels.predict}
\title{Submit model id and request a prediction}
\usage{
trainedmodels.predict(Input, id)
}
\arguments{
\item{Input}{The \link{Input} object to pass to this method}

\item{id}{The unique name for the predictive model}
}
\description{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_skeleton}}
}
\details{
Authentication scopes used by this function are:
\itemize{
  \item https://www.googleapis.com/auth/prediction
}

Set \code{options(googleAuthR.scopes.selected = c(https://www.googleapis.com/auth/prediction)}
Then run \code{googleAuthR::gar_auth()} to authenticate.
See \code{\link[googleAuthR]{gar_auth}} for details.
}
\seealso{
\href{https://developers.google.com/prediction/docs/developer-guide}{Google Documentation}

Other Input functions: \code{\link{Input.input}},
  \code{\link{Input}}, \code{\link{hostedmodels.predict}}
}
