% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prediction_objects.R
\name{Training.dataAnalysis}
\alias{Training.dataAnalysis}
\title{Training.dataAnalysis Object}
\usage{
Training.dataAnalysis(warnings = NULL)
}
\arguments{
\item{warnings}{No description}
}
\value{
Training.dataAnalysis object
}
\description{
Training.dataAnalysis Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Data Analysis.
}
\seealso{
Other Training functions: \code{\link{Training.modelInfo.confusionMatrixRowTotals}},
  \code{\link{Training.modelInfo.confusionMatrix}},
  \code{\link{Training.modelInfo}}, \code{\link{Training}},
  \code{\link{trainedmodels.insert}}
}
