% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prediction_objects.R
\name{Training}
\alias{Training}
\title{Training Object}
\usage{
Training(Training.dataAnalysis = NULL, Training.modelInfo = NULL,
  Training.modelInfo.confusionMatrix = NULL,
  Training.modelInfo.confusionMatrixRowTotals = NULL, dataAnalysis = NULL,
  id = NULL, modelInfo = NULL, selfLink = NULL,
  storageDataLocation = NULL, storagePMMLLocation = NULL,
  storagePMMLModelLocation = NULL, trainingStatus = NULL, utility = NULL)
}
\arguments{
\item{Training.dataAnalysis}{The \link{Training.dataAnalysis} object or list of objects}

\item{Training.modelInfo}{The \link{Training.modelInfo} object or list of objects}

\item{Training.modelInfo.confusionMatrix}{The \link{Training.modelInfo.confusionMatrix} object or list of objects}

\item{Training.modelInfo.confusionMatrixRowTotals}{The \link{Training.modelInfo.confusionMatrixRowTotals} object or list of objects}

\item{dataAnalysis}{Data Analysis}

\item{id}{The unique name for the predictive model}

\item{modelInfo}{Model metadata}

\item{selfLink}{A URL to re-request this resource}

\item{storageDataLocation}{Google storage location of the training data file}

\item{storagePMMLLocation}{Google storage location of the preprocessing pmml file}

\item{storagePMMLModelLocation}{Google storage location of the pmml model file}

\item{trainingStatus}{The current status of the training job}

\item{utility}{A class weighting function, which allows the importance weights for class labels to be specified [Categorical models only]}
}
\value{
Training object
}
\description{
Training Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
No description
}
\seealso{
Other Training functions: \code{\link{Training.dataAnalysis}},
  \code{\link{Training.modelInfo.confusionMatrixRowTotals}},
  \code{\link{Training.modelInfo.confusionMatrix}},
  \code{\link{Training.modelInfo}},
  \code{\link{trainedmodels.insert}}
}
